+++ objectified-ui/public/WHATS_NEW.md
@@ -36,6 +36,8 @@ We continue to improve the platform based on your feedback with improvements and
   - Canvas changes also apply to the Paths canvas
     - Added blur and background opacity settings
   - Search results now save history, and have filtering options for more complex search operations
+  - Export improvements:
+    - Added ability to export the entire viewport, or the current viewport
 - Paths Improvements:
   - Added API Key security schemes: define header, query, or cookie-based API keys in the Security tab and use them on operations (#410)
   - Added inline variable editor that can click parameters in the header of the path studio